+++ src/extensions/extension_globals.cpp
@@ -63,6 +63,7 @@ DynamicVectorClass<WarheadTypeClassExtension *> WarheadTypeExtensions;
 DynamicVectorClass<WaveClassExtension *> WaveExtensions;
 DynamicVectorClass<WeaponTypeClassExtension *> WeaponTypeExtensions;
 DynamicVectorClass<TEventClassExtension*> TEventExtensions;
+DynamicVectorClass<TActionClassExtension*> TActionExtensions;
 
 TacticalExtension *TacticalMapExtension = nullptr;